+++ CursedMod/Events/Arguments/SCPs/Scp079/PlayerUseLockdownEventArgs.cs
@@ -29,7 +29,7 @@ public class PlayerUseLockdownEventArgs : EventArgs, ICursedCancellableEvent, IC
 
     public CursedPlayer Player { get; }
     
-    public float PowerCost { get; set; }
+    public int PowerCost { get; set; }
     
     public CursedRoom Room { get; }
